+++ src/app/clusters/energy-preference-server/app_config_dependent_sources.cmake
@@ -16,6 +16,7 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/energy-preference-server.cpp"
+    "${CLUSTER_DIR}/CodegenIntegration.cpp"
+    "${CLUSTER_DIR}/energy-preference-delegate.h"
     "${CLUSTER_DIR}/energy-preference-server.h"
 )
