+++ src/app/clusters/energy-preference-server/app_config_dependent_sources.gni
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "energy-preference-server.cpp",
+  "CodegenIntegration.cpp",
+  "energy-preference-delegate.h",
   "energy-preference-server.h",
 ]
